refactor(tui): extract agent-config/change-watcher mixins from tui_gateway/server.py (shard s2)
