+++ tui_gateway/server.py
@@ -3176,278 +3176,6 @@ def _clear_pending(sid: str | None = None) -> None:
 # ── Agent factory ────────────────────────────────────────────────────
 
 
-def resolve_skin() -> dict:
-    try:
-        from hermes_cli.skin_engine import init_skin_from_config, get_active_skin
-
-        init_skin_from_config(_load_cfg())
-        skin = get_active_skin()
-        return {
-            "name": skin.name,
-            "colors": skin.colors,
-            # Paired palettes: the TUI detects the terminal's polarity and
-            # prefers the matching hand-tuned block over adapting `colors`.
-            "light_colors": skin.light_colors,
-            "dark_colors": skin.dark_colors,
-            "branding": skin.branding,
-            "banner_logo": skin.banner_logo,
-            "banner_hero": skin.banner_hero,
-            "tool_prefix": skin.tool_prefix,
-            "help_header": (skin.branding or {}).get("help_header", ""),
-        }
-    except Exception:
-        return {}
-
-
-# Signature of the last skin broadcast: (name, active user-file mtime). Lets the
-# per-tool reconcile fire ``skin.changed`` on any real move — a name switch OR a
-# live color edit to the active skin — and nothing else.
-_last_skin_sig: tuple[str, float | None] | None = None
-
-
-def _skin_sig() -> tuple[str, float | None]:
-    """(active skin name, its user-file mtime). Built-ins have no file, so only
-    their name moves; a user skin's mtime lets an in-place color edit repaint too."""
-    name = str((_load_cfg().get("display") or {}).get("skin") or "default")
-    override = get_hermes_home_override()
-    home = override if isinstance(override, str) and override else _hermes_home
-    try:
-        mtime: float | None = (Path(home) / "skins" / f"{name}.yaml").stat().st_mtime
-    except OSError:
-        mtime = None
-    return name, mtime
-
-
-def _note_skin_broadcast() -> None:
-    """Sync the reconcile baseline after the /skin RPC emits, so the per-tool
-    check doesn't re-broadcast the skin /skin just applied."""
-    global _last_skin_sig
-    try:
-        _last_skin_sig = _skin_sig()
-    except Exception:
-        pass
-
-
-def _broadcast_skin_if_changed() -> None:
-    """Emit ``skin.changed`` when the active skin moved — the agent switched it
-    (``hermes config set display.skin``) OR edited the active skin's colors in
-    place ("I don't like that coral" → tweak the YAML).
-
-    Routes through the SAME live path as ``/skin`` so every surface (TUI + desktop)
-    repaints, no slash command. The signature check is a dict lookup + one stat,
-    so polling it is ~free.
-    """
-    global _last_skin_sig
-    try:
-        sig = _skin_sig()
-    except Exception:
-        return
-    if sig == _last_skin_sig:
-        return
-    _last_skin_sig = sig
-    try:
-        _broadcast_global_event("skin.changed", resolve_skin())
-    except Exception:
-        pass
-
-
-def _watcher_home() -> Path:
-    """Active profile home for the change watcher's signature probes."""
-    override = get_hermes_home_override()
-    return Path(override if isinstance(override, str) and override else _hermes_home)
-
-
-def _pet_sig() -> tuple:
-    """(slug, spritesheet revision, scale) of the active pet — ("off",) when none.
-
-    Cheap by construction: config comes from the mtime-cached ``_load_cfg`` and
-    the sheet revision is one stat. Moves when ``/pet`` (de)activates a pet, the
-    hatch flow rebuilds a sheet, or the scale changes."""
-    display = _load_cfg().get("display") or {}
-    pet_cfg = display.get("pet") if isinstance(display.get("pet"), dict) else {}
-    if not pet_cfg or not pet_cfg.get("enabled"):
-        return ("off",)
-    try:
-        enabled, pet, scale = _pet_active_selection()
-        if not enabled or pet is None or not pet.exists:
-            return ("off",)
-        return (pet.slug, _pet_sheet_revision(pet.spritesheet), scale)
-    except Exception:  # noqa: BLE001 - cosmetic, never break the watcher
-        return ("off",)
-
-
-def _pet_changed_payload() -> dict:
-    """``pet.info.meta``-shaped payload for ``pet.changed`` — enough for the
-    renderer to decide whether the heavy sprite payload needs a refetch."""
-    try:
-        enabled, pet, scale = _pet_active_selection()
-        if not enabled or pet is None or not pet.exists:
-            return {"enabled": False}
-        return {
-            "enabled": True,
-            "slug": pet.slug,
-            "displayName": pet.display_name,
-            "scale": scale,
-            "spritesheetRevision": _pet_sheet_revision(pet.spritesheet),
-        }
-    except Exception:  # noqa: BLE001 - cosmetic, never break the watcher
-        return {"enabled": False}
-
-
-def _cron_sig():
-    """mtime of the profile's cron/jobs.json — moves on create/edit/pause/
-    remove AND on scheduler tick bookkeeping (last_run/next_run)."""
-    try:
-        return (_watcher_home() / "cron" / "jobs.json").stat().st_mtime_ns
-    except OSError:
-        return None
-
-
-def _sessions_sig():
-    """Newest mtime across state.db and its WAL — the cross-process change
-    signal. Messaging-gateway turns and cron runs are written by OTHER
-    processes that never touch this gateway's transports; the shared SQLite
-    file is the one thing they all move (#58671)."""
-    home = _watcher_home()
-    sig = None
-    for name in ("state.db", "state.db-wal"):
-        try:
-            mtime = (home / name).stat().st_mtime_ns
-        except OSError:
-            continue
-        sig = mtime if sig is None else max(sig, mtime)
-    return sig
-
-
-def _platforms_sig():
-    """mtime of gateway_state.json — the messaging gateway process persists
-    platform connect/disconnect/health there, so its movement is the
-    "connection status changed" signal for the Messaging page."""
-    try:
-        return (_watcher_home() / "gateway_state.json").stat().st_mtime_ns
-    except OSError:
-        return None
-
-
-def _pairing_sig():
-    """Newest mtime across every profile's pairing store.
-
-    An unknown DMer's pending code is written by the messaging gateway — a
-    DIFFERENT process that never touches this gateway's transports — so the
-    files are the only shared signal. ``platforms.changed`` cannot stand in
-    for this: it tracks connect/disconnect/health, and a pairing request
-    moves nothing in gateway_state.json.
-    """
-    home = _watcher_home()
-    sig = None
-    # Global store (legacy `pairing/` and consolidated `platforms/pairing/`)
-    # plus every named profile's own — the Messaging page can be scoped to any
-    # of them, and a request landing in a profile store must still tick.
-    roots = [home / "pairing", home / "platforms" / "pairing"]
-    try:
-        for profile_dir in (home / "profiles").iterdir():
-            roots.append(profile_dir / "pairing")
-            roots.append(profile_dir / "platforms" / "pairing")
-    except OSError:
-        pass
-
-    for root in roots:
-        try:
-            entries = list(root.iterdir())
-        except OSError:
-            continue
-        for entry in entries:
-            # Only the pending/approved ledgers — _rate_limits.json moves on
-            # every unauthorized DM, including ones that produce no new row.
-            if not entry.name.endswith(("-pending.json", "-approved.json")):
-                continue
-            try:
-                mtime = entry.stat().st_mtime_ns
-            except OSError:
-                continue
-            sig = mtime if sig is None else max(sig, mtime)
-    return sig
-
-
-# Watched change signals: event → (check interval, signature fn, payload fn).
-# Signatures are stat/dict-lookup cheap, same bar as the skin watcher; the
-# check interval keeps the pricier probes (pet resolves the active sheet off
-# disk) off the 0.5s tick.
-_CHANGE_WATCHES: dict[str, tuple[float, Any, Any]] = {
-    "pet.changed": (2.0, _pet_sig, _pet_changed_payload),
-    "cron.changed": (1.0, _cron_sig, lambda: {}),
-    "sessions.changed": (0.5, _sessions_sig, lambda: {}),
-    "platforms.changed": (2.0, _platforms_sig, lambda: {}),
-    "pairing.changed": (2.0, _pairing_sig, lambda: {}),
-}
-
-# state.db moves on every message append during a streaming turn, and the
-# gateway rewrites gateway_state.json for in-flight-count bookkeeping; the
-# floor coalesces those bursts to one broadcast per window (trailing edge
-# included — a floored change keeps its old signature and re-fires next tick).
-_CHANGE_BROADCAST_FLOOR_S = {"sessions.changed": 2.0, "platforms.changed": 5.0}
-
-_change_sigs: dict[str, Any] = {}
-_change_checked_at: dict[str, float] = {}
-_change_broadcast_at: dict[str, float] = {}
-
-
-def _broadcast_watched_changes(now: float | None = None) -> None:
-    """One pass over ``_CHANGE_WATCHES``: recompute due signatures, broadcast
-    the events whose signature moved. First sighting seeds silently so a
-    gateway boot never fires a spurious refresh storm."""
-    now = time.monotonic() if now is None else now
-    for event, (interval, sig_fn, payload_fn) in _CHANGE_WATCHES.items():
-        if now - _change_checked_at.get(event, -interval) < interval:
-            continue
-        _change_checked_at[event] = now
-        try:
-            sig = sig_fn()
-        except Exception:  # noqa: BLE001 - a broken probe must not kill the loop
-            continue
-        if event not in _change_sigs:
-            _change_sigs[event] = sig
-            continue
-        if sig == _change_sigs[event]:
-            continue
-        floor = _CHANGE_BROADCAST_FLOOR_S.get(event, 0.0)
-        if floor and now - _change_broadcast_at.get(event, -floor) < floor:
-            # Floored: leave the old signature in place so the change re-fires
-            # once the window opens (the trailing edge of the burst).
-            continue
-        _change_sigs[event] = sig
-        _change_broadcast_at[event] = now
-        try:
-            _broadcast_global_event(event, payload_fn())
-        except Exception:  # noqa: BLE001
-            pass
-
-
-_skin_watcher_started = False
-
-
-def _ensure_skin_watcher() -> None:
-    """Watch cheap on-disk signatures and broadcast change events — so a skin
-    Hermes activates, a pet ``/pet`` adopts, a cron the scheduler fires, or a
-    messaging turn another process writes goes live on every surface within a
-    couple seconds, on its own, with no client-side poll in the loop.
-    Idempotent; started at gateway.ready. (Named for its original skin-only
-    duty; it is the process's one change watcher.)"""
-    global _skin_watcher_started
-    if _skin_watcher_started:
-        return
-    _skin_watcher_started = True
-    _note_skin_broadcast()  # seed the baseline so only a real change repaints
-
-    def _loop() -> None:
-        while True:
-            time.sleep(0.5)
-            _broadcast_skin_if_changed()
-            _broadcast_watched_changes()
-
-    threading.Thread(target=_loop, name="hermes-change-watcher", daemon=True).start()
-
-
 def _resolve_model() -> str:
     env = (
         os.environ.get("HERMES_MODEL", "")
@@ -3895,313 +3623,6 @@ def _write_config_key(key_path: str, value):
         current = current[key]
     current[keys[-1]] = value
     _save_cfg(cfg)
-
-
-_STATUSBAR_MODES = frozenset({"off", "top", "bottom"})
-_APPROVAL_MODES = frozenset({"manual", "smart", "off"})
-
-
-def _load_approval_mode() -> str:
-    """Resolve the effective ``approvals.mode`` for the TUI surface.
-
-    Delegates to the canonical resolver in ``tools.approval``
-    (``_get_approval_mode``) so mode resolution cannot drift per surface —
-    the same normalization, defaults, and config precedence the approval
-    gate itself uses (see ``tools/approval.py``).
-
-    Previously this re-read the config raw via ``_load_cfg`` +
-    ``_deep_merge(DEFAULT_CONFIG, ...)`` and normalized locally, which
-    could disagree with the gate's own view of the mode (e.g. the
-    canonical ``hermes_cli.config.load_config`` path applies managed-scope
-    overlays and ``${VAR}`` env expansion that the TUI's raw YAML read did
-    not fully mirror).
-    """
-    from tools.approval import _get_approval_mode
-
-    mode = _get_approval_mode()
-    return mode if mode in _APPROVAL_MODES else "manual"
-
-
-def _coerce_statusbar(raw) -> str:
-    if raw is False:
-        return "off"
-    if isinstance(raw, str) and (s := raw.strip().lower()) in _STATUSBAR_MODES:
-        return s
-    return "top"
-
-
-_MOUSE_TRACKING_ALIASES = {
-    "0": "off",
-    "1": "all",
-    "all": "all",
-    "any": "all",
-    "button": "buttons",
-    "buttons": "buttons",
-    "click": "buttons",
-    "false": "off",
-    "full": "all",
-    "no": "off",
-    "off": "off",
-    "on": "all",
-    "scroll": "wheel",
-    "true": "all",
-    "wheel": "wheel",
-    "yes": "all",
-}
-
-
-def _display_mouse_tracking(display: dict) -> str:
-    """Resolve display.mouse_tracking to one of ``off|wheel|buttons|all``.
-
-    Boolean values keep their legacy meaning (``True`` → ``all``, ``False`` →
-    ``off``). The ``wheel`` preset (DEC 1000+1006) is the tmux-friendly
-    subset — wheel + click only, no hover events to trigger prompt-row
-    clipboard probes. Legacy ``tui_mouse`` is honored only when
-    ``mouse_tracking`` is absent.
-    """
-    if not isinstance(display, dict):
-        return "all"
-    if "mouse_tracking" in display:
-        raw = display.get("mouse_tracking")
-    else:
-        raw = display.get("tui_mouse", True)
-    if raw is False or raw == 0:
-        return "off"
-    if raw is True or raw is None:
-        return "all"
-    if isinstance(raw, (int, float)):
-        return "all"
-    if isinstance(raw, str):
-        return _MOUSE_TRACKING_ALIASES.get(raw.strip().lower(), "all")
-    return "all"
-
-
-def _load_reasoning_config(model: str = "") -> dict | None:
-    """Load reasoning effort from config.yaml, respecting per-model overrides.
-
-    Thin wrapper over the shared chokepoint
-    :func:`hermes_constants.resolve_reasoning_config` (per-model override >
-    global ``agent.reasoning_effort``; YAML boolean False = disabled).
-    Closes #21256.
-    """
-    from hermes_constants import resolve_reasoning_config
-
-    return resolve_reasoning_config(_load_cfg(), model)
-
-
-def _load_service_tier() -> str | None:
-    raw = (
-        str((_load_cfg().get("agent") or {}).get("service_tier", "") or "")
-        .strip()
-        .lower()
-    )
-    if not raw or raw in {"normal", "default", "standard", "off", "none"}:
-        return None
-    if raw in {"fast", "priority", "on"}:
-        return "priority"
-    return None
-
-
-def _load_provider_routing() -> dict:
-    """OpenRouter provider-routing prefs from config.yaml (``provider_routing``).
-
-    Parity with the messaging gateway (``gateway/run.py::_load_provider_routing``)
-    and the classic CLI: without this the desktop/TUI backend builds agents with
-    no routing prefs, so OpenRouter falls back to its default (effectively random)
-    provider selection even when the user configured ``provider_routing``.
-    """
-    try:
-        return _load_cfg().get("provider_routing", {}) or {}
-    except Exception:
-        return {}
-
-
-def _load_show_reasoning() -> bool:
-    # Fallback True — keep in sync with DEFAULT_CONFIG display.show_reasoning
-    # (this loader reads the raw user YAML without the DEFAULT_CONFIG merge).
-    return bool((_load_cfg().get("display") or {}).get("show_reasoning", True))
-
-
-def _load_memory_notifications() -> str:
-    """Self-improvement review notification mode from config.yaml.
-
-    Parity with the messaging gateway (``gateway/run.py``) and the classic CLI:
-    ``display.memory_notifications`` controls whether the background review's
-    "💾 Self-improvement review: …" summary is surfaced. Without this the
-    TUI/desktop backend always behaved as ``"on"`` and silently ignored a user
-    who set ``off``. Accepts ``off`` / ``on`` (default) / ``verbose``; a bool is
-    normalized for back-compat.
-    """
-    raw = (_load_cfg().get("display") or {}).get("memory_notifications")
-    if isinstance(raw, bool):
-        return "on" if raw else "off"
-    return str(raw).lower() if raw else "on"
-
-
-def _load_tool_progress_mode() -> str:
-    env = os.environ.get("HERMES_TUI_TOOL_PROGRESS", "").strip().lower()
-    if env in {"off", "new", "all", "verbose"}:
-        return env
-    raw = (_load_cfg().get("display") or {}).get("tool_progress", "all")
-    if raw is False:
-        return "off"
-    if raw is True:
-        return "all"
-    mode = str(raw or "all").strip().lower()
-    return mode if mode in {"off", "new", "all", "verbose"} else "all"
-
-
-def _load_enabled_toolsets() -> list[str] | None:
-    explicit = [
-        item.strip()
-        for item in os.environ.get("HERMES_TUI_TOOLSETS", "").split(",")
-        if item.strip()
-    ]
-    cfg = None
-    fallback_notice = None
-
-    # Coding posture (base Hermes): with no explicit pin, collapse to the
-    # coding toolset (+ enabled MCP servers) when sitting in a code workspace.
-    # The desktop app and `hermes --tui` both land here. See
-    # agent/coding_context.py. No config is loaded yet at this point, so we let
-    # coding_selection() load it lazily (cli.py passes its already-resolved
-    # CLI_CONFIG instead, purely to avoid a redundant read).
-    if not explicit:
-        try:
-            from agent.coding_context import coding_selection
-
-            selection = coding_selection(platform=_resolve_session_platform())
-            if selection is not None:
-                # Fold in `project` here too: this is a GUI-only resolver, and
-                # the focus-mode coding posture returns before the fallback path
-                # that normally adds it — without this the desktop loses the
-                # project tools exactly when sitting in a repo (see below).
-                return sorted({*selection, "project"})
-        except Exception:
-            pass
-
-    try:
-        from toolsets import validate_toolset
-    except Exception:
-        validate_toolset = None
-
-    if explicit and validate_toolset is not None:
-        built_in = [name for name in explicit if validate_toolset(name)]
-        unresolved = [name for name in explicit if name not in built_in]
-
-        if unresolved:
-            try:
-                from hermes_cli.plugins import discover_plugins
-
-                discover_plugins()
-                plugin_valid = [name for name in unresolved if validate_toolset(name)]
-            except Exception:
-                plugin_valid = []
-
-            if plugin_valid:
-                built_in.extend(plugin_valid)
-                unresolved = [name for name in unresolved if name not in plugin_valid]
-
-        if any(name in {"all", "*"} for name in built_in):
-            ignored = [name for name in explicit if name not in {"all", "*"}]
-            if ignored:
-                print(
-                    "[tui] HERMES_TUI_TOOLSETS=all enables every toolset; "
-                    f"ignoring additional entries: {', '.join(ignored)}",
-                    file=sys.stderr,
-                    flush=True,
-                )
-            return None
-
-        if not unresolved:
-            return built_in
-
-        mcp_names: set[str] = set()
-        mcp_disabled: set[str] = set()
-        try:
-            from hermes_cli.config import read_raw_config
-            from hermes_cli.tools_config import _parse_enabled_flag
-
-            raw_cfg = read_raw_config()
-            mcp_servers = (
-                raw_cfg.get("mcp_servers")
-                if isinstance(raw_cfg.get("mcp_servers"), dict)
-                else {}
-            )
-            for name, server_cfg in mcp_servers.items():
-                if not isinstance(server_cfg, dict):
-                    continue
-                if _parse_enabled_flag(server_cfg.get("enabled", True), default=True):
-                    mcp_names.add(str(name))
-                else:
-                    mcp_disabled.add(str(name))
-        except Exception:
-            mcp_names = set()
-            mcp_disabled = set()
-
-        mcp_valid = [name for name in unresolved if name in mcp_names]
-        disabled = [name for name in unresolved if name in mcp_disabled]
-        unknown = [
-            name
-            for name in unresolved
-            if name not in mcp_names and name not in mcp_disabled
-        ]
-        valid = built_in + mcp_valid
-
-        if unknown:
-            print(
-                f"[tui] ignoring unknown HERMES_TUI_TOOLSETS entries: {', '.join(unknown)}",
-                file=sys.stderr,
-                flush=True,
-            )
-        if disabled:
-            print(
-                "[tui] ignoring disabled MCP servers in HERMES_TUI_TOOLSETS "
-                "(set enabled: true in config.yaml to use): "
-                f"{', '.join(disabled)}",
-                file=sys.stderr,
-                flush=True,
-            )
-
-        if valid:
-            return valid
-
-        fallback_notice = (
-            "[tui] no valid HERMES_TUI_TOOLSETS entries; using configured CLI toolsets"
-        )
-
-    try:
-        from hermes_cli.config import load_config
-        from hermes_cli.tools_config import _get_platform_tools
-
-        cfg = cfg if cfg is not None else load_config()
-
-        # Runtime toolset resolution must include default MCP servers so the
-        # agent can actually call them. Passing ``False`` here is the
-        # config-editing variant — used when we need to persist a toolset
-        # list without baking in implicit MCP defaults. Using the wrong
-        # variant at agent creation time makes MCP tools silently missing
-        # from the TUI. See PR #3252 for the original design split.
-        enabled = _get_platform_tools(cfg, "cli", include_default_mcp_servers=True)
-        if fallback_notice is not None:
-            print(fallback_notice, file=sys.stderr, flush=True)
-        if not enabled:
-            return None
-        # The desktop Project tools are off _HERMES_CORE_TOOLS (every other
-        # platform would carry their schema for nothing), so the platform
-        # recovery above — which keys off hermes-cli's tool universe — can't
-        # surface them. This resolver runs ONLY in the desktop/TUI gateway, so
-        # folding in the `project` toolset here is the gate that exposes them on
-        # exactly the surface that can follow a project move.
-        return sorted(enabled | {"project"})
-    except Exception:
-        if fallback_notice is not None:
-            print(
-                "[tui] no valid HERMES_TUI_TOOLSETS entries and configured CLI toolsets could not be loaded; enabling all toolsets",
-                file=sys.stderr,
-                flush=True,
-            )
-        return None
 
 
 def _session_tool_progress_mode(sid: str) -> str:
@@ -13901,5 +13322,18 @@ for _m in (
     _methods_complete,
     _methods_tools,
 ):
+    _m.register(sys.modules[__name__])
+del _m
+
+
+# ── Wave-1 split modules (see method_ctx.py) ─────────────────────────
+# Imported at the end of this module so every global the moved functions
+# close over already exists; register() rebinds them onto this namespace.
+from . import (  # noqa: E402
+    agent_config_loaders as _agent_config_loaders,
+    change_watcher_mixin as _change_watcher_mixin,
+)
+
+for _m in (_agent_config_loaders, _change_watcher_mixin):
     _m.register(sys.modules[__name__])
 del _m
